Updating the problem of difference between strings and atoi(0)
When I used atoi to code, then I realised that I couldn't make the difference
between 0 and non strinc-characters, like atoi('a'). They all give 0, so I created
a function name containString that return -1, when there is a string character, so
not in the interval [0, 9]
And, now, I can make the difference!